Convert interrupts to a power of two modulus division
C compilers typically convert modulus of powers of two to the equivalent
bitwise and, which makes those operations a fair amount faster.

Compare the assembly generated by the two previous and new values
https://godbolt.org/g/1LAov4

This is admittedly minor, but as this check needs to occur every loop
iteration it can add up.