Forward caller env in partialised functions
This is meant to support lexically scoped methods from the definition
environment of the partialised function.

This helps with this case:

```
local({
  mean.foobar <- function(...) "foobar"
  foobar <- structure(list(), class = "foobar")

  mean(foobar) == partial(mean)(foobar)
})
```

That's not standard dispatch semantics. Ideally we'd dispatch in the
caller environment rather than the definition environment. The issue
is that there's a fundamental conflict between these goals:

(a) Evaluating arguments in their environment (typically def env)
(b) Allowing substitution of partialised arguments
(c) Lexical dispatch in caller env rather than def env

It might just be that partialised functions are meant to be private or
even anonymous (and thus local). Also lexical dispatch in the global
env should work anyway because most envs inherit from the search
path. And if in a package, registration will take care of dispatch.
Let's not worry about this too much.